arm64: insn: Don't fallback on nosync path for general insn patching
Patching kernel instructions at runtime requires other CPUs to undergo
a context synchronisation event via an explicit ISB or an IPI in order
to ensure that the new instructions are visible. This is required even
for "hotpatch" instructions such as NOP and BL, so avoid optimising in
this case and always go via stop_machine() when performing general
patching.

ftrace isn't quite as strict, so it can continue to call the nosync
code directly.

Change-Id: I5f9e73209afd263a83b8f7f107503809c8d97019
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>